Split program into separate files
This commit moves all the ptrace-related code into its own file, and
also moves the capability-acquiring code into another file. This way
it will be easier to add other methods of waiting in the future.
